Add files via upload
This is a rough draft data management plan (DMP) that needs a lot of work and discussion with team members to more fully reflect the project. The project does not officially begin until January, 2020. The DMP attached only describes my portion of a much larger research and will be modified extensively.